+++ internal/internal_task_pollers.go
@@ -123,6 +123,7 @@ type (
 		pendingStickyPollCount  int
 		stickyBacklog           int64
 		requestLock             sync.Mutex
+		slotAdmission           *workflowSlotAdmission
 		stickyCacheSize         int
 		eagerActivityExecutor   *eagerActivityExecutor
 
@@ -1239,6 +1240,14 @@ func (wtp *workflowTaskPoller) updateBacklog(taskQueueKind enumspb.TaskQueueKind
 	wtp.requestLock.Lock()
 	wtp.stickyBacklog = backlogCountHint
 	wtp.requestLock.Unlock()
+
+	if wtp.slotAdmission != nil {
+		wtp.slotAdmission.setStickyBacklog(backlogCountHint)
+	}
+}
+
+func (wtp *workflowTaskPoller) setSlotAdmission(admission *workflowSlotAdmission) {
+	wtp.slotAdmission = admission
 }
 
 // getNextPollRequest returns appropriate next poll request based on poller configuration and mode.